+++ R/import_toxswa.R
@@ -4,273 +4,167 @@
 #' Read all TOXSWA files within a SWASH project directory.
 #'
 #' @param swash_dir path to the SWASH project directory
+#' @param ... arguments passed on to [import_toxswa()]
 #'
-#' @return a list of exposure profiles each with two columns (time and concentration)
+#' @return a list of imported exposure series, see [import_toxswa()] for details
 #' @export
-import_swash <- function(swash_dir){
+import_swash <- function(swash_dir, ...){
   toxwa_file_regex <- "^\\d*\\.out$"
   f <- list.files(swash_dir,
                   pattern = toxwa_file_regex,
                   full.names = TRUE)
-  import_toxswa(f)
+  import_toxswa(f, ...)
 }
 
-#' Read TOXSWA exposure profiles from out-files
+#' Import *TOXSWA* exposure series
 #'
-#' Read several TOXSWA exposure profiles from out-files at their corresponding
-#' paths given by'pathtofiles'.
+#' Read one or more *TOXSWA* exposure series from *TOXSWA*'s `.out` files. By default, the
+#' concentration dissolved in water (*ConLiqWatLay*) at the end of the simulated waterbody
+#' (i.e. at the maximum of the *x* dimension) is returned. The unit of the time
+#' scale as well as of the imported model output variable can be scaled as needed.
 #'
-#' @param pathtofiles paths to toxswa files that should be read
-#' @param profileNames a vector of names to be used for the loaded toxswa files. if "NA" the filename without extension will be used
+#' The numerical time scale is shifted to always start at time zero (0.0).
+#' Numerical columns of the returned *data.frame* objects will be of type
+#' [units::units]. Please be aware that the use of `units` objects may not be
+#' supported by all functions in this package. However, [set_times()] and
+#' [set_exposure()] can handle `units` objects safely.
 #'
+#' Incomplete list of alternative *TOXSWA* v5.5.3 output variables:
+#' - *ConLiqWatLay*: Concentration dissolved in water (g/m3)
+#' - *ConLiqSed*: Concentration in pore water sediment (g/m3)
+#' - *ConSysWatLay*: Total concentration in water (g/m3)
+#' - *CntSorSusSol*: Content sorbed to suspended solids (g/kg)
+#' - *CntSorSed*: Content sorbed to sediment (g/kg)
+#'
+#' @param filepath vector of strings with absolute or relative paths to files
+#' @param output_var character, single output variable from *TOXSWA* that is
+#'   imported, defaults to *ConLiqWatLay*
+#' @param output_unit character, target unit of the imported output variable,
+#'   defaults to *ug/L*, syntax must be compatible with [units::units()]
+#' @param time_unit character, target unit of the imported time scale,
+#'   defaults to *days*, syntax must be compatible with [units::units()]
+#' @param substance optional vector of characters, if set, only the substance
+#'   codes defined in this vector are imported
+#' @param split logical, if `TRUE` then one series will be returned for each
+#'   substance found in the *TOXSWA* files, else all substances per file will
+#'   be in one *data.frame*. Defaults to `FALSE`
 #' @export
+#' @autoglobal
+#' @return list of *data.frame* objects with exposure series. Each *data.frame* has at
+#'   least three columns:
+#'   - `time`: numerical time scale, always starts at zero
+#'   - `timestamp`: time as datetime objects such as `POSIXct`
+#'   - one or more additional columns for each imported substance
 #'
-#' @return list of exposure profiles each with two columns (time and concentration)
-import_toxswa <- function(pathtofiles, profileNames = NA){
-  if ( any(sapply(profileNames, is.na)) ) {
-    profileNames <- basename(pathtofiles) %>%
-      tools::file_path_sans_ext()
-  } else {
-    profileNames <- profileNames
+import_toxswa <- function(filepath, output_var="ConLiqWatLay", output_unit="ug/L",
+                           time_unit="days", substance=NULL, split=FALSE) {
+  if(length(filepath) == 0) {
+    stop("filepath vector is empty")
   }
-  profiles <- lapply(pathtofiles, extract_focus_profile)
-  names(profiles) <- profileNames
-
-  flag_null <- do.call(c, lapply(profiles, is.null))
-  profiles <- profiles[!flag_null]
-
-  profiles <- lapply(setNames(names(profiles), names(profiles)),
-                     function(x){
-                       split_toxswa_exposure_profile(profiles[[x]], x)
-                     })
-  profiles <- flatten_list(profiles)
-
-  profiles <- lapply(profiles, melt_focus_profiles)
-
-  #trialnames <- rep_len(profileNames,2)
-  profiles <- lapply(setNames(names(profiles),names(profiles)), function(i){
-    profiles[[i]]["trial"] <- i
-    profiles[[i]]
-  })
-
-  return(profiles)
-}
-
-#' Extract a focus profile from a single .out file
-#'
-#' Extract the relevant information (i.e. concentration time series) from a TOXSWA out-file.
-#'
-#' @param file_name the name of the out-file from which the data is extracted
-#' @param layer_string the TOXSWA layer whose data is extracted
-#' @param substance_string the name of the substance
-#'
-#' @return exposure profiles
-#' @global V1 V2 V3 DateTime
-#' @noRd
-extract_focus_profile <- function(
-    file_name,
-    layer_string = "ConLiqWatLay",
-    substance_string = ""
-) {
-
-  lin <- readLines(file_name, encoding = "UTF-8")
-
-  isToxswa <- check_if_toxswa(lin)
-  if ( !isToxswa ){
-    message("file is not a toxswa file")
-    return(NULL)
+  if(length(output_var) != 1)
+    stop("Exactly one output variable must be selected")
+  # vectorized file list
+  if(length(filepath) > 1) {
+    return(sapply(filepath, import_toxswa, output_var=output_var, output_unit=output_unit,
+                  time_unit=time_unit, substance=substance, split=split, USE.NAMES=FALSE))
   }
 
-  is_supported_version <- check_if_toxswa_version(lin, supported_version = c("4","5.5.3"))
-  if ( !is_supported_version ){
-    message(paste0("Toxswa version is not supported. So far only version 4 and 5.5.3 are supported."))
-    return(NULL)
+  if(any(is.na(filepath)) | !is.character(filepath))
+    stop("filepath vector contains invalid values")
+  if(!file.exists(filepath))
+    stop("file does not exists: ", filepath)
+
+  headers <- paste0("V", seq(1, max(utils::count.fields(filepath, sep=""))))
+  df <- utils::read.table(filepath, header=FALSE, col.names=headers, fill=TRUE)
+
+  comments <- df %>%
+    dplyr::filter(V1=="*") %>%
+    tidyr::unite(col="V1", dplyr::everything(), sep=" ", na.rm=TRUE) %>%
+    dplyr::mutate(V1=trimws(V1)) %>%
+    dplyr::pull(V1)
+  data <- df %>%
+    dplyr::filter(V1 != "*")
+  rm(df, headers)
+
+  # check toxswa version
+  version_string <- comments[stringr::str_detect(comments, "TOXSWA version")]
+  if(length(version_string) == 0)
+    stop("TOXSWA version missing")
+  if(!stringr::str_detect(version_string, " (4|5\\.5\\.3)$"))
+    stop("TOXSWA version not supported")
+
+  # get unit of chosen output var
+  unit_str <- comments[stringr::str_detect(comments, paste0("^\\* Unit for ", output_var, " is "))]
+  if(length(unit_str) == 0)
+    stop("Unit of output var missing")
+  if(length(unit_str) > 1)
+    stop("Output variable is not unique")
+  unit_str <- stringr::str_extract(unit_str, "\\((.*?)\\)", group=1)
+  if(is.na(unit_str) | unit_str == "")
+    stop("Unit of output var missing")
+  # replace any '.' by a space to avoid issues with `units` package
+  unit_str <- stringr::str_replace(unit_str, "\\.", " ")
+
+  # is output var included in timerseries?
+  data <- data %>%
+    dplyr::filter(stringr::str_detect(tolower(V3), pattern=paste0("^", tolower(output_var), "(_\\w+)?$"))) %>%
+    dplyr::select(dplyr::where(~ sum(!is.na(.x)) > 0)) %>% # remove any columns containing only NA
+    dplyr::select(dplyr::where(~ sum(.x != "") > 0)) # remove any columns containing only emtpy strings
+  if(nrow(data) == 0)
+    stop("No data for output variable in TOXSWA file")
+
+  data <- data %>%
+    dplyr::select(c(dplyr::any_of(c(1, 2, 3)), dplyr::last_col())) %>%
+    dplyr::rename(time=1, timestamp=2, name=3, value=4) %>%
+    dplyr::mutate(value=units::set_units(as.numeric(value), unit_str, mode="standard"))
+
+  # convert units of output if requested
+  if(!is.null(output_unit)) {
+    units(data$value) <- units::as_units(output_unit)
   }
 
-  unit <- lin %>%
-    stringr::str_subset(
-      paste0("\\* Unit for ", head(stringr::str_split(layer_string, "_")[[1]],1), " is")
-    ) %>%
-    stringr::str_split("[\\(\\)]") %>%
-    unlist() %>%
-    magrittr::extract(2) %>%
-    stringr::str_trim() %>%
-    stringr::str_replace_all("\\.", " ")
+  # check if variable is substance-dependent (which it should be), and
+  # pivot by substance
+  if(any(stringr::str_detect(data$name, "_")))
+  {
+    data <- data %>%
+      dplyr::mutate(name=stringr::str_extract(name, "_(.+)$", group=1)) %>%
+      tidyr::pivot_wider()
+  }
+  else if(!is.null(substance)) {
+    stop("Output var is not substance dependent, check inputs")
+  }
 
-  lin %>%
-    stringr::str_subset(stringr::str_c(layer_string, "_", substance_string)) %>%
-    stringr::str_split(" +") %>%
-    sapply(function(x) c(x[3:4], tail(x,1))) %>%
-    t() %>%
-    dplyr::as_tibble(.name_repair = ~paste0("V", seq_along(.))) %>%
-    dplyr::transmute(
-      Layer_Specifier = as.character(V2),
-      DateTime = as.character(V1) %>% lubridate::dmy_hm(),
-      `Time` = units::set_units(
-        (DateTime - DateTime[1]) / lubridate::ddays(1),
-        "d",
-        mode = "standard"),
-      Exposure = units::set_units(as.numeric(V3), unit, mode = "standard")
-    ) %>%
-    tibble::add_column(
-      RunID =
-        stringr::str_split(file_name, "[/\\\\]")[[1]] %>%
-        tail(1) %>%
-        stringr::str_extract("[[:digit:]]+"),
-      .before = "DateTime"
-    ) %>%
-    dplyr::select(dplyr::all_of(c("RunID", "Layer_Specifier", "DateTime", "Time", "Exposure"))) %>%
-    dplyr::rename_at("Exposure", ~ paste0("Concentration"))
-}
+  # is chosen substance part of result?
+  if(!is.null(substance)) {
+    if(!all(substance %in% names(data)))
+      stop("Selected substance code not found in file")
 
-#' Split a profile into a list if more then one substances was detected
-#'
-#' @param profile the toxswa profile as a tibble
-#'
-#' @return a tibble or a list of tibbles if more than one substance was identified
-#' @noRd
-#' @examples
-#' dat <- data.frame(RunID = 123,
-#'   Layer_Specifier = rep(c("ConLiqWatLay_ai1","ConLiqWatLay_ai2"),10),
-#'   Concentration = rlnorm(20,0,1)
-#'   )
-#'   split_toxswa_exposure_profile(dat)
-split_toxswa_exposure_profile <- function(profile, runid=NA){
+    data <- data %>% dplyr::select(c(1, 2, !!substance))
+  }
 
-  # ai holds the name of the active ingredient for each point in the time series
-  ai <- sub("^ConLiqWatLay_(.+)$","\\1",profile[["Layer_Specifier"]])
+  data <- data %>%
+    dplyr::mutate(time=as.numeric(time),
+                  time=time - time[[1]],
+                  time=units::set_units(time, "days"),
+                  time=units::set_units(time, time_unit, mode="standard"),
+                  timestamp=lubridate::dmy_hm(timestamp))
+  # set unit information as attribute for backwards compat
+  myunits <- c("time"=units::deparse_unit(data$time),
+               "conc"=units::deparse_unit(as.data.frame(data)[,3]))
+  attr(data, "units") <- myunits
 
-  if (length(unique(ai)) == 1 ){ # if only one ai return the profile
-    return(profile)
-
-  } else if (length(unique(ai)) > 1){ # if more than one split the time series in a list
-    if (is.na(runid)){
-      runid_ <- profile[1,"RunID"]
-    } else {
-      runid_ <- runid
+  fn <- tools::file_path_sans_ext(basename(filepath))
+  result <- list()
+  # split series by substance?
+  if(split) {
+    nms <- names(data)
+    for(i in seq(3, length(data))) {
+      key <- paste0(fn, "_", nms[[i]])
+      result[[key]] <- dplyr::select(data, dplyr::all_of(c(1, 2, i)))
     }
-    out <- split(profile, profile[,"Layer_Specifier"])
-    names(out) <- paste0(runid_,"_",unique(ai))
-    return(out)
-
   } else {
-    return(NULL)
+    result[[fn]] <- data
   }
-
+  result
 }
-
-
-#' Checks if the input contains the default TOXSWA keywords in the header
-#'
-#' @param ln
-#'
-#' @return TRUE if header matches the default TOXSWA header
-#' @noRd
-#' @examples
-#'  ln <- c("TOXSWA REPORT: Header",
-#'  "* Results from the TOXSWA model  (c) Alterra",
-#'  "* FOCUS  TOXSWA version   : 4",
-#'  "* TOXSWA model version    : 3.3.4",
-#'  "* TOXSWA created on       : 02-Apr-2015")
-#'  check_if_toxswa(ln)
-check_if_toxswa <- function(ln){
-  ln <- ln[1:20]
-  strings_to_match <- c(
-    "^\\* Results from the TOXSWA model\\s*\\(c\\)",
-    "^\\* FOCUS\\s*TOXSWA version\\s*:",
-    "^\\* TOXSWA model version\\s*:",
-    "^\\* TOXSWA created on\\s*:"
-  )
-
-  contains_strings <- do.call(c,lapply(strings_to_match, function(x) any(grepl(pattern=x, ln))))
-
-  !any(!contains_strings)
-
-}
-
-
-#' Reduce the focus exposure profile
-#'
-#' Reduce to a two column data.frame with first column indicating the time
-#' and the second column the concentration
-#'
-#' @param profile the profile data
-#' @param time_col the name of the column with the time data
-#' @param conc_col the name of the column with the concentration data
-#'
-#' @return the reduced FOCUS profile
-#' @noRd
-melt_focus_profiles <- function(profile, time_col = "Time", conc_col =  "Concentration"){
-  out <- profile %>%
-    dplyr::select(dplyr::all_of(c(time_col,conc_col))) %>%
-    units::drop_units() %>%
-    as.data.frame()
-
-  names(out) <- c("time", "conc")
-
-  myunits <- setNames(
-    c(profile %>% dplyr::pull(time_col) %>% units::deparse_unit(),
-      profile %>% dplyr::pull(conc_col) %>% units::deparse_unit()),
-    names(out)
-  )
-  attr(out, "units") <- myunits
-
-  return(out)
-}
-
-
-#' Checks if the toxswa version is as specified
-#'
-#' @param ln a vector of strings
-#' @param supported_version a string of the version that is looked for
-#'
-#' @return TRUE if version numbers match
-#'
-#' @noRd
-#' @examples
-#' ln <- c("* ------------------------------------------------------------------------------",
-#' "* Results from the TOXSWA model  (c) Wageningen University & Research",
-#' "* FOCUS  TOXSWA version   : 5.5.3",
-#' "* TOXSWA model version    : 3.3.6",
-#' "* TOXSWA created on       : 17-Dec-2017")
-#' check_if_toxswa_version(ln, supported_version = "5.5.3")
-#'
-#'  ln2 <- c("TOXSWA REPORT: Header",
-#'  "* Results from the TOXSWA model  (c) Alterra",
-#'  "* FOCUS  TOXSWA version   : 4",
-#'  "* TOXSWA model version    : 3.3.4",
-#'  "* TOXSWA created on       : 02-Apr-2015")
-#' check_if_toxswa_version(ln2)
-check_if_toxswa_version <- function(ln, supported_version = c("4", "5.5.3") ){
-  regex_version <- "^\\*\\s+FOCUS\\s+TOXSWA\\s+version"
-  Toxswa_version_line <- grep(regex_version, ln, value = TRUE)
-  Toxswa_version <- sub(paste0(regex_version,"\\s+:\\s*(.*)$"),"\\1", Toxswa_version_line)
-  Toxswa_version %in% supported_version
-}
-
-#' Flatten a list to first layer
-#'
-#' @param lst the input list
-#'
-#' @return a list
-#'
-#' @noRd
-#' @examples
-#' iris_list <- split(iris, iris$Species)
-#' nested_list <- list(iris_list[1], list(iris_list[2], iris_list[3]))
-#' nested_list
-#' flatten_list(nested_list)
-flatten_list <- function(lst) {
-  flattened_list <- list()
-
-  for (i in seq_along(lst)) {
-    if ( !("list" %in% class(lst[[i]])) ) {
-      flattened_list <- c(flattened_list, lst[i])
-    } else {
-      nested_list <- Recall(lst[[i]])
-      flattened_list <- c(flattened_list, nested_list)
-    }
-  }
-  return(flattened_list)
-}
-